#458 Restructure component modifiers to allow modification of non-annotated methods
