+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/core/services/ServiceModifier.java
@@ -17,22 +17,18 @@
 
 package org.dockbox.hartshorn.core.services;
 
+import org.dockbox.hartshorn.core.ComponentType;
 import org.dockbox.hartshorn.core.context.ApplicationContext;
 import org.dockbox.hartshorn.core.context.element.TypeContext;
-import org.dockbox.hartshorn.core.inject.InjectionModifier;
 import org.dockbox.hartshorn.core.properties.Attribute;
 import org.jetbrains.annotations.Nullable;
 
 import java.lang.annotation.Annotation;
 
-public abstract class ServiceModifier<A extends Annotation> implements InjectionModifier<A> {
+public abstract class ServiceModifier<A extends Annotation> extends ComponentModifier<A> {
 
     @Override
     public <T> boolean preconditions(final ApplicationContext context, final TypeContext<T> type, @Nullable final T instance, final Attribute<?>... properties) {
-        return context.locator().container(type).present()
-                && this.modifies(context, type, instance, properties);
+        return super.preconditions(context, type, instance, properties) && context.locator().container(type).get().componentType() == ComponentType.FUNCTIONAL;
     }
-
-    protected abstract <T> boolean modifies(ApplicationContext context, TypeContext<T> type, @Nullable T instance, Attribute<?>... properties);
-
 }
